+++ src/app/static/src/app/store/adr/actions.ts
@@ -25,7 +25,7 @@ export interface ADRActions {
     getSchemas: (store: ActionContext<ADRState, RootState>) => void;
     getUserCanUpload: (store: ActionContext<ADRState, RootState>) => void;
     getUploadFiles: (store: ActionContext<ADRState, RootState>) => void;
-    uploadFilestoADR: (store: ActionContext<ADRState, RootState>, uploadFilesPayload: UploadFile[]) => void;
+    uploadFilesToADR: (store: ActionContext<ADRState, RootState>, uploadFilesPayload: UploadFile[]) => void;
 }
 
 export const actions: ActionTree<ADRState, RootState> & ADRActions = {
@@ -76,30 +76,10 @@ export const actions: ActionTree<ADRState, RootState> & ADRActions = {
         const selectedDataset = rootState.baseline.selectedDataset;
 
         if (selectedDataset) {
-            //For backward compatibility, we may have to regenerate the dataset metadata to provide the
-            //organisation id for projects which are reloaded
-            let selectedDatasetOrgId: string;
             if (!selectedDataset.organization) {
-                //We may also have to fetch the selected dataset metadata too, if not loaded during this session
-                let datasets = state.datasets;
-                if (!datasets.length) {
-                    await api(context)
-                        .ignoreErrors()
-                        .ignoreSuccess()
-                        .get(`/adr/datasets/${selectedDataset.id}`)
-                        .then((response) => {
-                            if (response) {
-                                datasets = [response.data];
-                            }
-                        });
-                }
-
-                const regenDataset = datasetFromMetadata(selectedDataset.id, datasets, state.schemas!);
-                commit(`baseline/${BaselineMutation.SetDataset}`, regenDataset, {root: true});
-                selectedDatasetOrgId = regenDataset.organization.id;
-            } else {
-                selectedDatasetOrgId = selectedDataset.organization.id;
+                await getAndSetDatasets(context, selectedDataset.id)
             }
+            const selectedDatasetOrgId = rootState.baseline.selectedDataset!.organization.id
 
             await api(context)
                 .withError(ADRMutation.SetADRError)
@@ -188,10 +168,12 @@ export const actions: ActionTree<ADRState, RootState> & ADRActions = {
         }
     },
 
-    async uploadFilestoADR(context, uploadFilesPayload) {
-        const {state, rootState, commit} = context;
-        const selectedDatasetId = rootState.baseline.selectedDataset?.id;
+    async uploadFilesToADR(context, uploadFilesPayload) {
+        const {state, rootState, commit, dispatch} = context;
+        const uploadMetadata = rootState.modelRun.result?.uploadMetadata
+        const selectedDatasetId = rootState.baseline.selectedDataset!.id;
         const modelCalibrateId = rootState.modelCalibrate.calibrateId;
+
         commit({type: ADRMutation.ADRUploadStarted, payload: uploadFilesPayload.length});
 
         for (let i = 0; i < uploadFilesPayload.length; i++) {
@@ -199,8 +181,18 @@ export const actions: ActionTree<ADRState, RootState> & ADRActions = {
             const { resourceType, resourceFilename, resourceName, resourceId } = uploadFilesPayload[i];
 
             const requestParams: Dict<string> = {resourceFileName: resourceFilename, resourceName};
-            if (resourceId){
+            if (resourceId) {
                 requestParams["resourceId"] = resourceId
+            }
+            if (resourceType === state.schemas?.outputSummary) {
+                requestParams["description"] = uploadMetadata
+                    ? uploadMetadata.outputSummary.description
+                    : "Naomi summary report uploaded from Naomi web app"
+            }
+            if (resourceType === state.schemas?.outputZip) {
+                requestParams["description"] = uploadMetadata
+                    ? uploadMetadata.outputZip.description
+                    : "Naomi output uploaded from Naomi web app"
             }
 
             let apiRequest = api<ADRMutation, ADRMutation>(context)
@@ -216,6 +208,27 @@ export const actions: ActionTree<ADRState, RootState> & ADRActions = {
                 break
             }
         }
+        await getAndSetDatasets(context, selectedDatasetId)
+        dispatch("getUploadFiles");
     }
 };
+
+async function getAndSetDatasets(context: ActionContext<ADRState, RootState>, selectedDatasetId: string){
+    const {state, commit} = context;
+    let datasets = state.datasets;
+    if (!datasets.length) {
+        await api(context)
+                .ignoreErrors()
+                .ignoreSuccess()
+                .get(`/adr/datasets/${selectedDatasetId}`)
+                .then((response) => {
+                    if (response) {
+                        datasets = [response.data];
+                    }
+                });
+    }
+
+    const regenDataset = datasetFromMetadata(selectedDatasetId, datasets, state.schemas!);
+    commit(`baseline/${BaselineMutation.SetDataset}`, regenDataset, {root: true});
+}