+++ src/app/static/src/app/store/modelCalibrate/actions.ts
@@ -84,7 +84,6 @@ export const actions: ActionTree<ModelCalibrateState, RootState> & ModelCalibrat
                         },
                         {root: true});
                 }
-
                 commit(ModelCalibrateMutation.Calibrated);
             }
         }