+++ src/app/static/src/tests/components/downloadResults/uploadModal.test.ts
@@ -6,6 +6,7 @@ import {mockADRState, mockBaselineState, mockDatasetResource} from "../../mocks"
 import registerTranslations from "../../../app/store/translations/registerTranslations";
 import {expectTranslated} from "../../testHelpers";
 import Vue from 'vue';
+import { Dict } from "../../../app/types";
 
 describe(`uploadModal `, () => {
 
@@ -45,7 +46,7 @@ describe(`uploadModal `, () => {
     const mockOrganization = {
         id: "123abc"
     }
-    const createStore = (data = fakeMetadata) => {
+    const createStore = (data: Dict<any> = fakeMetadata) => {
         const store = new Vuex.Store({
             state: emptyState(),
             modules: {
@@ -67,7 +68,7 @@ describe(`uploadModal `, () => {
                     namespaced: true,
                     state: mockADRState({uploadFiles: data}),
                     actions: {
-                        uploadFilestoADR: jest.fn()
+                        uploadFilesToADR: jest.fn()
                     },
                     mutations: {
                         ADRUploadStarted: jest.fn()
@@ -122,8 +123,9 @@ describe(`uploadModal `, () => {
 
         const overwriteText = wrapper.findAll("small")
         expect(overwriteText.length).toBe(1)
-        expect(overwriteText.at(0).text()).toBe("This file already exists on ADR " +
-            "and will be overwritten. File was updated 25/01/2021 06:34:12")
+        expectTranslated(overwriteText.at(0), "This file already exists on ADR " +
+            "and will be overwritten. File was updated 25/01/2021 06:34:12",
+            "Ce fichier existe déjà sur ADR et sera écrasé. Le fichier a été mis à jour 25/01/2021 06:34:12", store)
 
         const inputs = wrapper.findAll("input.form-check-input")
         expect(inputs.length).toBe(2)
@@ -136,8 +138,10 @@ describe(`uploadModal `, () => {
     })
 
     it(`checkboxes are set by default`, async () => {
-        const wrapper = shallowMount(UploadModal, {store: createStore()})
-        await wrapper.setData({uploadFilesToAdr: ["outputZip", "outputSummary"]})
+        const store = createStore({})
+        const wrapper = shallowMount(UploadModal, {store})
+        store.state.adr.uploadFiles = fakeMetadata
+        await Vue.nextTick()
         const inputs = wrapper.findAll("input.form-check-input")
         expect(inputs.length).toBe(2)
 
@@ -157,18 +161,6 @@ describe(`uploadModal `, () => {
 
         expect(wrapper.find("#dialog").exists()).toBe(true)
         expect(wrapper.vm.$data.uploadFilesToAdr).toMatchObject(["outputZip", "outputSummary"])
-    })
-
-    it(`can set and get description value as expected`, async () => {
-        const wrapper = getWrapper()
-        const newDesc = "new description"
-
-        const desc = wrapper.find("#description-id")
-        desc.setValue(newDesc)
-        expect(wrapper.vm.$data.uploadDescToAdr).toBe(newDesc)
-
-        const textarea = desc.element as HTMLTextAreaElement
-        expect(textarea.value).toBe(newDesc)
     })
 
     it(`can trigger close modal as expected`, async () => {