+++ src/app/static/src/tests/integration/adr-dataset.itest.ts
@@ -7,7 +7,6 @@ import {SurveyAndProgramMutation} from "../../app/store/surveyAndProgram/mutatio
 import {getFormData} from "./helpers";
 import {ADRMutation} from "../../app/store/adr/mutations";
 import {UploadFile} from "../../app/types";
-import {mockADRState} from "../mocks";
 
 // this suite tests all endpoints that talk to the ADR
 // we put them in a suite of their own so that we can run
@@ -104,7 +103,11 @@ describe("ADR dataset-related actions", () => {
     });
 
     it("can get dataset details on get userCanUpload when selected dataset organisation is not set", async () => {
-        const commit = jest.fn();
+        const commit = jest.fn().mockImplementation((mutation, payload) => {
+            if (mutation === "baseline/SetDataset") {
+                root.baseline.selectedDataset = payload
+            }
+        });
 
         // 1. get datasets
         await adrActions.getDatasets({commit, rootState} as any);
@@ -208,6 +211,7 @@ describe("ADR dataset-related actions", () => {
     it("hits upload files to adr endpoint and gets appropriate error", async () => {
 
         const commit = jest.fn();
+        const dispatch = jest.fn();
         const root = {
             ...rootState,
             modelCalibrate: {calibrateId: "calId"},
@@ -215,6 +219,14 @@ describe("ADR dataset-related actions", () => {
                 selectedDataset: {
                     id: "datasetId",
                     organization: {id: "organisationId"}
+                }
+            },
+            modelRun: {
+                result: {
+                    uploadMetadata: {
+                        outputSummary: {description: "summary"},
+                        outputZip: {description: "zip"}
+                    }
                 }
             }
         };
@@ -229,16 +241,19 @@ describe("ADR dataset-related actions", () => {
             }
         ] as UploadFile[]
 
-        await adrActions.uploadFilestoADR({commit, state: adr, rootState: root} as any,
-            uploadFilesPayload);
+        await adrActions.uploadFilesToADR({commit, dispatch, state: adr, rootState: root} as any, uploadFilesPayload);
 
-        expect(commit.mock.calls.length).toBe(3);
+        expect(commit.mock.calls.length).toBe(4);
         expect(commit.mock.calls[0][0]["type"]).toBe("ADRUploadStarted");
         expect(commit.mock.calls[0][0]["payload"]).toBe(1);
         expect(commit.mock.calls[1][0]["type"]).toBe("ADRUploadProgress");
         expect(commit.mock.calls[1][0]["payload"]).toBe(1);
         expect(commit.mock.calls[2][0]["type"]).toBe("SetADRUploadError");
         expect(commit.mock.calls[2][0]["payload"]["error"]).toBe("OTHER_ERROR");
+        expect(commit.mock.calls[3][0]).toBe("baseline/SetDataset");
+        expect(commit.mock.calls[3][2]["root"]).toBe(true);
+        expect(dispatch.mock.calls.length).toBe(1);
+        expect(dispatch.mock.calls[0][0]).toBe("getUploadFiles");
     }, 7000);
 
 });